@@ -127,7 +127,7 @@ class EditImagePage extends ConsumerWidget {
               borderRadius: BorderRadius.circular(7),
               boxShadow: [
                 BoxShadow(
-                  color: Colors.black.withOpacity(0.2),
+                  color: Colors.black.withValues(alpha: 0.2),
                   spreadRadius: 2,
                   blurRadius: 10,
                   offset: const Offset(0, 3),
