@@ -202,12 +202,12 @@ class ThumbnailImage extends ConsumerWidget {
                   bottom: 5,
                   child: Icon(
                     storageIcon(asset),
-                    color: Colors.white.withOpacity(.8),
+                    color: Colors.white.withValues(alpha: .8),
                     size: 16,
                     shadows: [
                       Shadow(
                         blurRadius: 5.0,
-                        color: Colors.black.withOpacity(0.6),
+                        color: Colors.black.withValues(alpha: 0.6),
                         offset: const Offset(0.0, 0.0),
                       ),
                     ],
