@@ -27,7 +27,8 @@ class ScaffoldErrorBody extends StatelessWidget {
               child: Icon(
                 Icons.error_outline,
                 size: 100,
-                color: context.themeData.iconTheme.color?.withOpacity(0.5),
+                color:
+                    context.themeData.iconTheme.color?.withValues(alpha: 0.5),
               ),
             ),
           ),
