@@ -126,7 +126,7 @@ class _BlurredBackdrop extends HookWidget {
           ),
         ),
         child: Container(
-          color: Colors.black.withOpacity(0.2),
+          color: Colors.black.withValues(alpha: 0.2),
         ),
       );
     } else {
@@ -147,7 +147,7 @@ class _BlurredBackdrop extends HookWidget {
             ),
           ),
           child: Container(
-            color: Colors.black.withOpacity(0.2),
+            color: Colors.black.withValues(alpha: 0.2),
           ),
         ),
       );
