@@ -22,7 +22,7 @@ class SearchFilterChip extends StatelessWidget {
         onTap: onTap,
         child: Card(
           elevation: 0,
-          color: context.primaryColor.withOpacity(.5),
+          color: context.primaryColor.withValues(alpha: .5),
           shape: StadiumBorder(
             side: BorderSide(color: context.colorScheme.secondaryContainer),
           ),
